New hubInfoHandler method to replace obsolete hubUpdatedHandler (#99)
* Update webcore.groovy

fixing bugs with hubPowerSource.  this module was not listening for the correct event (was listening for HubUpdated instead of hubInfo), so it was not properly generating a value change for hubPowerSource.  that has been fixed and now the value change works as intended.

* Update webcore.groovy

fixing tab vs space

* Update webcore.groovy

* Update webcore.groovy

* Update webcore.groovy

adding legacy event subscription information and associated method handler back in
adding back in intentional ? in method call